add confirmation before deleting user from either list

+++ src/Components/AllUsers.js
@@ -5,12 +5,17 @@ import { removeGenericUserAction, removeSpecificUserAction } from '../Store/Acti
 
 export default function AllUsers ({ users, header, genericOrSpecific }) {
   const dispatch = useDispatch();
+  const confirmString = 'Are you sure you want to delete this user from the';
 
   const removeUserHandler = (name) => {
     switch (genericOrSpecific) {
       case 'generic':
+        const confirmDelGeneral = window.confirm(`${confirmString} general list`);
+        if (!confirmDelGeneral) return;
         return dispatch(removeGenericUserAction(name));
       case 'specific':
+        const confirmDelSpecific = window.confirm(`${confirmString} pecific list`);
+        if (!confirmDelSpecific) return;
         return dispatch(removeSpecificUserAction(name));
       default:
         return;
